Move `AgentAutoRegistrationProperties` into the common module.
