+++ agent/src/com/thoughtworks/go/agent/service/SslInfrastructureService.java
@@ -17,6 +17,7 @@
 
 package com.thoughtworks.go.agent.service;
 
+import com.thoughtworks.go.config.AgentAutoRegistrationProperties;
 import com.thoughtworks.go.config.AgentRegistry;
 import com.thoughtworks.go.config.GuidService;
 import com.thoughtworks.go.security.AuthSSLProtocolSocketFactory;
